pyproject.toml
[project]
name = "dbtui"
version = "0.1.0"
description = "A terminal UI for dbt (data build tool)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-artifacts-parser>=0.12.0",
    "pyyaml>=6.0.3",
    "textual>=7.3.0",
]

[project.scripts]
dbtui = "dbtui.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
